Plop logic into a trait, since PHP 7.3 still wasn't cool with us typehinting throwable
