+++ src/Exceptions/ControlPanelExceptionHandlerForLaravelSix.php
@@ -7,8 +7,10 @@ use Illuminate\Foundation\Exceptions\Handler;
 
 class ControlPanelExceptionHandlerForLaravelSix extends Handler
 {
+    use Concerns\RendersControlPanelExceptions;
+
     public function render($request, Exception $e)
     {
-        return app(ControlPanelExceptionHandler::class)->render($request, $e);
+        return $this->renderException($request, $e);
     }
 }
